fix memory fragmentation

+++ kmr.cpp
@@ -107,7 +107,7 @@ uint32_t round_pow2(uint32_t n, bool up) {
 	return lo;
 }
 
-uint32_t *deserialize(char *fn, uint32_t *len) {
+uint32_t *deserialize(char *fn, uint64_t *len) {
 	ifstream ifs(fn);
 	ifs.seekg(0, ifs.end);
 	*len = ifs.tellg();
